pyproject.toml
[project]
name = "recettesbot"
version = "0.1.0"
description = "RecettesBot est un bot Telegram qui permet aux d'enregistrer des recettes à partir de n'importe quelle URL, de les rechercher par titre grâce à la recherche vectorielle, d'exporter les recettes au format PDF, d'ajuster le nombre de portions et de générer une liste de courses via un sondage interactif."
authors = [
    { name = "j-about", email = "142051449+j-about@users.noreply.github.com" }
]
requires-python = ">=3.13.13"
dependencies = [
    "alembic==1.18.4",
    "asyncpg==0.31.0",
    "claude-agent-sdk==0.1.58",
    "fpdf2==2.8.7",
    "logfire==4.31.1",
    "pgvector==0.4.2",
    "psycopg2-binary==2.9.11",
    "pydantic-settings==2.13.1",
    "python-telegram-bot==22.7",
    "sentence-transformers==5.4.0",
    "sqlmodel==0.0.38",
]

[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "ruff==0.15.10",
    "ty==0.0.29",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
python-version = "3.13"

[tool.ty.rules]
# Start permissive; tighten as the codebase grows.

# pydantic-settings BaseSettings populates required fields from environment
# variables, but ty doesn't model that yet — it sees the constructor as missing
# arguments. Suppress only for the modules that instantiate Settings().
[[tool.ty.overrides]]
include = ["settings.py", "tests/test_settings.py"]

[tool.ty.overrides.rules]
missing-argument = "ignore"

# SQLAlchemy's `ischema_names` is typed as a dict of a closed union of built-in
# dialect types, so registering a third-party type (pgvector's VECTOR) trips
# `invalid-assignment`. The assignment is the documented pgvector pattern; ty
# can't model it without a stub update.
[[tool.ty.overrides]]
include = ["migrations/env.py"]

[tool.ty.overrides.rules]
invalid-assignment = "ignore"

# SQLModel's metaclass adds `__table__` at runtime that ty can't see, and
# `Field(sa_type=...)` is documented to accept type instances even though the
# overload signature wants `type[Any]`. Column-vs-value comparisons in
# `where(...)` clauses also surface as bool to ty because SQLModel doesn't
# expose the SQLAlchemy descriptor protocol that would type them as
# `ColumnElement[bool]`. Both patterns dominate the model test suites.
[[tool.ty.overrides]]
include = [
    "bot.py",
    "tests/e2e/conftest.py",
    "tests/e2e/test_flows.py",
    "tests/integration/test_database.py",
    "tests/integration/test_models.py",
    "tests/mocks/stubs.py",
    "tests/mocks/factories.py",
    "tests/unit/test_bot.py",
    "tests/unit/test_models.py",
    "tests/unit/test_pdf.py",
]

[tool.ty.overrides.rules]
invalid-argument-type = "ignore"
unresolved-attribute = "ignore"
# `Ingredient(name=..., quantity=..., unit=...)` inside a `Recipe(ingredients=[...])`
# initializer has no `recipe_id` because SQLModel sets it at flush time via the
# relationship — ty cannot model that.
missing-argument = "ignore"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=0"
